feat : create connection to mysql, utilize flyway for database migration, create stored procedures, and add test for controller layer

+++ src/main/java/nl/fontys/s3/huister/business/interfaces/city/GetAllCitiesUseCase.java
@@ -3,5 +3,5 @@ package nl.fontys.s3.huister.business.interfaces.city;
 import nl.fontys.s3.huister.business.response.city.GetAllCitiesResponse;
 
 public interface GetAllCitiesUseCase {
-    GetAllCitiesResponse getAllCities(int userId);
+    GetAllCitiesResponse getAllCities(long userId);
 }
