+++ src/main/java/nl/fontys/s3/huister/business/interfaces/property/DeletePropertyUseCase.java
@@ -1,4 +1,4 @@
 package nl.fontys.s3.huister.business.interfaces.property;
 public interface DeletePropertyUseCase {
-    void deleteProperty(final int id);
+    void deleteProperty(final long id);
 }
